Start extracting a react-kle-keymap library
The goal here is to have all the hardware libraries that have their keymaps
based on a KLE-provided SVG be rebuilt on top of a library that makes it
considerably simpler to lay things out.

It calculates the x, y offsets, all one has to worry about is key positions, not
pixel positions. Key positions are considerably easier to wrap one's head
around.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>